fix(dashboard): close invitation-email enumeration and tighten sentry replay masking
- requireEmailMatch was a public Server Action callable directly over
  the network, returning distinguishable errors for valid-token and
  email-mismatch cases. Move to lib/auth/invitation.ts (no use server
  directive); throw a single generic Invalid invitation for both paths.
- Sentry replay defaults: maskAllText=true and blockAllMedia=true so
  multi-tenant text in dashboard pages is not captured by default.
  Per-selector mask list retained as defense in depth.
